KOGITO-7373 Zanini review
